fix: don't overwrite user config changes with default theme
The default theme was loaded after the user configuration file was read,
which led to user changes being overwritten by default values.  Reverse
the order of operations so that the default theme is loaded first and
the user configuration file is read afterwards to modify the existing
theme or to replace it outright.

Drop support for `${STEAMSHIP_THEME}` variable. To set the theme from
the configuration file, simply run `steamship theme <name>` to load the
named theme.